Use a date picker in Search
This looks simple, but that DatePicker component is ridiculous if you try to use
custom inputs. There's like 5 different issues in their repo of people trying to
figure out how to get it working, and nobody ever seems to actually get there.
This isn't really relevant for this commit, but I spent an ungodly amount of
time making this work and need to vent.

Nobody reads these anyway. I'm shouting into the void.